requirements.txt
titiler-core>=1.2.0
uvicorn>=0.41.0
dhis2eo @ git+https://github.com/dhis2/dhis2eo.git@v1.1.0
dhis2-client @ git+https://github.com/dhis2/dhis2-python-client.git@V0.3.0
earthkit-transforms==0.5.*
metpy>=1.7,<2